# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commercial Dining Intelligence API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/merchant/predict": {
      "post": {
        "summary": "Merchant Predict",
        "operationId": "merchant_predict_api_v1_merchant_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantPredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantPredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/heatmap": {
      "get": {
        "summary": "Merchant Heatmap Get",
        "description": "Browser-friendly probe; heavy work uses POST.",
        "operationId": "merchant_heatmap_get_api_v1_merchant_heatmap_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Merchant Heatmap Get Api V1 Merchant Heatmap Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Merchant Heatmap",
        "operationId": "merchant_heatmap_api_v1_merchant_heatmap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantHeatmapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantHeatmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/cities": {
      "get": {
        "summary": "Merchant Cities",
        "operationId": "merchant_cities_api_v1_merchant_cities_get",
        "parameters": [
          {
            "name": "min_rows",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 50,
              "title": "Min Rows"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantCitiesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/categories": {
      "get": {
        "summary": "Merchant Categories",
        "operationId": "merchant_categories_api_v1_merchant_categories_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "max_rows_if_no_city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 100,
              "default": 2000,
              "title": "Max Rows If No City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantCategoriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/categories/resolve": {
      "get": {
        "summary": "Merchant Categories Resolve",
        "operationId": "merchant_categories_resolve_api_v1_merchant_categories_resolve_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "description": "Free text, e.g. 'burger', 'fast food, coffee' — matched to train_spatial cat_* in this city slice.",
              "title": "Q"
            },
            "description": "Free text, e.g. 'burger', 'fast food, coffee' — matched to train_spatial cat_* in this city slice."
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "max_rows_if_no_city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 100,
              "default": 2000,
              "title": "Max Rows If No City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantCategoriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/coverage": {
      "get": {
        "summary": "Merchant Coverage",
        "operationId": "merchant_coverage_api_v1_merchant_coverage_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "max_rows_if_no_city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 100,
              "default": 2000,
              "title": "Max Rows If No City"
            }
          },
          {
            "name": "max_sample_points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 50,
              "default": 400,
              "title": "Max Sample Points"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantCoverageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search": {
      "post": {
        "summary": "Search V1",
        "operationId": "search_v1_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MerchantCategoriesResponse": {
        "properties": {
          "category_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Category Keys",
            "description": "cat_* names in train_spatial that appear at least once in the slice (all-zero columns dropped unless all are zero)."
          }
        },
        "type": "object",
        "title": "MerchantCategoriesResponse",
        "description": "cat_* names aligned with the predict slice for the current city/state/max_rows."
      },
      "MerchantCitiesResponse": {
        "properties": {
          "cities": {
            "items": {
              "$ref": "#/components/schemas/MerchantCityRow"
            },
            "type": "array",
            "title": "Cities"
          }
        },
        "type": "object",
        "required": [
          "cities"
        ],
        "title": "MerchantCitiesResponse"
      },
...
      "MerchantHeatmapRequest": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "Optional USPS state (must match train_spatial.state when used)."
          },
          "category_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Category Keys",
            "description": "Explicit cat_* columns; optional if category_query is set."
          },
          "category_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Query",
            "description": "Plain text resolved to cat_* in the current slice (same as /merchant/predict)."
          },
          "max_rows_if_no_city": {
            "type": "integer",
            "maximum": 50000.0,
            "minimum": 100.0,
            "title": "Max Rows If No City",
            "default": 2000
          },
          "price_level": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Level",
            "description": "Yelp-style tier vs local mean within 1 km."
          },
          "price_per_person": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Person"
          },
          "grid_size": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 4.0,
            "title": "Grid Size",
            "description": "Number of rows/columns over min/max lat/lon of the reference slice (max 256 cells).",
            "default": 12
          }
        },
        "type": "object",
        "title": "MerchantHeatmapRequest",
        "description": "Same filters as merchant predict, without a pin — scores a regular grid over the slice bbox."
      },
...
      "MerchantPredictRequest": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City",
            "description": "Filter reference businesses; if empty, use the first N rows of the table"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "Optional USPS state to disambiguate duplicate city names (must match train_spatial.state)."
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "category_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Category Keys",
            "description": "Explicit train_spatial cat_* columns. Optional if category_query is set."
          },
          "category_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Query",
            "description": "Plain text (e.g. 'fast food, coffee', 'burger'); server maps to cat_* in the current slice."
          },
          "max_rows_if_no_city": {
            "type": "integer",
            "maximum": 50000.0,
            "minimum": 100.0,
            "title": "Max Rows If No City",
            "default": 2000
          },
          "price_level": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Level",
            "description": "Yelp-style price tier 1–4; compared to mean attr_restaurantspricerange2 within 1 km."
          },
          "price_per_person": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Person",
            "description": "USD-ish per person; mapped to 1–4 if price_level not set."
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "MerchantPredictRequest"
      },
      "MerchantPredictResponse": {
        "properties": {
          "survival_probability": {
            "type": "number",
            "title": "Survival Probability"
          },
          "predicted_stars": {
            "type": "number",
            "title": "Predicted Stars"
          },
          "reference_row_count": {
            "type": "integer",
            "title": "Reference Row Count"
          },
          "city_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Filter"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          },
          "live_feature_preview": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Live Feature Preview"
          },
          "inside_reference_hull": {
            "type": "boolean",
            "title": "Inside Reference Hull",
            "description": "True if the pin lies inside the convex hull of reference training coordinates for this slice."
          },
          "resolved_category_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Resolved Category Keys",
            "description": "cat_* columns used for this prediction (from category_query and/or category_keys)."
          },
          "price_fit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Fit",
            "description": "good | medium | poor — rule-based match vs mean local price level (1 km)."
          },
          "price_gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Gap"
          },
          "nearby_avg_price_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nearby Avg Price Level"
          },
          "risk": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Risk",
            "description": "competition, location, price: high|medium|low|unknown (rule-based)."
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "description": "Short human-readable summary for the business user.",
            "default": ""
          },
          "business_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Score",
            "description": "0–100 composite from survival, stars, competition, price match (rule-based V1)."
          },
          "business_score_ml": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Score Ml",
            "description": "0–100 supervised score: P(is_open) from spatial/category features only (no survival/rating heads as inputs). Null if artifact missing."
          }
        },
        "type": "object",
        "required": [
          "survival_probability",
          "predicted_stars",
          "reference_row_count",
          "inside_reference_hull"
        ],
        "title": "MerchantPredictResponse"
      },
...
    }
  }
}